// src/utils/Type.ts
export type field = {
  series: string;
  match: string;
  time: string;
  invesment: number;
  returns: number;
  numberOfPlayer: number;
  created: string;
  comment: string;
  date: string;
};

export type dropdown = {
  id: number;
  name: number | string;
  value: number | string;
};

// export type create = {
//   id: number;
//   value: string | number;
//   name: string;
// };